LYSS Read + Platforms
The outer layer that you made is not in-line with the angle of rest of the structure so it looks kinda weird when platforming. I wanted to fix this and so I changed your whole bottom wall to be able to be turned from the center of the structure. There is something else that we need to do but I'm still thinking a bout a solution for that.